requirements.txt
streamlit
akshare
baostock
pandas
pytesseract
Pillow